Move asset views
Move asset views to item file, move collection asset views to collection file.
