+++ app/stac_api/views/views_test.py
@@ -3,10 +3,10 @@ import logging
 from rest_framework import generics
 
 from stac_api.models import LandingPage
+from stac_api.views.collection import CollectionAssetDetail
 from stac_api.views.collection import CollectionDetail
+from stac_api.views.item import AssetDetail
 from stac_api.views.item import ItemDetail
-from stac_api.views.views import AssetDetail
-from stac_api.views.views import CollectionAssetDetail
 
 logger = logging.getLogger(__name__)
